Continued room ownership and some Bug fixes
Began adding to the ManInfoWindow class to get an interface for locating and transferring the guests' Owned Rooms. This is incomplete at the moment, will continue to work on this.

Bug Fix: Book Guest Button does not require two attempts to open anymore. I still don't know exactly what was causing this, but I found a work around that seems to be unique to this particular window.

Bug Fix: Guests and Cleaners don't deteriorate and clean the rooms their assigned to until they are stationed inside of the room (before they would affect the room as soon as they start walking to it)