Guest Bedroom Ownership Complete
Needs polishing.

Bedrooms can keep track of their owners, regardless of where they are in the hotel. Similarly, guests can keep track of their rooms.

Place new guest with no room ownership in a bedroom with a vacant owner slot to assign them as an owner.

Click on a guest to open their InfoWindow. If they own a room, the room's name, View button, and Change button will be visible.

View highlights the room with a selector.

Change sets the game's state to a "ChangeRoomState" that allows the user to click on a room to attempt to replace the currently owned bedroom with the clicked on bedroom. Right click to return to the normal ManInfoWindow state. *Needs a visual representation on if a room has a vacant Owner slot before clicking on it*