Pushing the Cleaner Closet Room
Also pushed a quick system for roles that workers can have, as well as their payment. Currently the payment is just deducted, no functionality for it not being payable, or a change in their behavior if it is not payed.
Pushing some changes to the scene to include the new room prefab in the roommanager, as well as a small change to the GameManager inspector for the roles functionality