Unit Animation Swapping Bugfix
Fixed Unit animations bug with pathfinding. The animations will unfortunately just be jittery for a while, until the state system within the manscript is updated. The animations need to be changed on different times, as of right now the unit would attempt to change animations multiple times within a frame and end up not being able to switch to the proper animation.