Pushing Guest Room Assignment Check Change
Did some refactoring to the rear-end code organization and now the find closest bedroom makes sure the bedroom is within the cleanliness threshhold. Also made the cleaners check every now and then if a room has become clean by subscribing to a "room cleaned event" that I added to the Worker class that is called any time that they finish cleaning a room.