Dirty Particles in rooms
Got the particles implemented to change color depending on the current room cleanliness. Particles turn on once the room becomes less than 80% clean and become 40% green at the exact point of the room reaching 0% cleanliness. Added a new material that uses the shader in order to instantiate instances of it to separate changing the float value that makes the dirty particles become green.